fix(reports): a surface states the population it measures
runAdHocPlan builds a spec and hands it to the same runSpec the catalog
reports use, so the population landed on the tool surface too — and a
team-scoped reader's project aggregate went to zero. A project is read
by every seat holding the object grant, and an aggregate that narrowed
it would tell a delivery lead their department has no projects running.

Population is a property of the SURFACE, not of the engine, so the spec
now carries it. measureCallersOwn is the zero value on purpose: a report
added to the catalog that says nothing measures the caller's own, so the
narrow answer is what omission gives you and widening is written down.

report.go and reportsql.go each crossed the 500-line cap on the way, so
the ad-hoc plan and the WHERE side take the two files their names were
already asking for.
