+++ backend/internal/compose/report.go
@@ -180,6 +180,10 @@ type reportSpec struct {
 	filters      map[string]string
 	defaultBy    []string
 	defaultAggs  []reportAggregate
+	// population says whose records this report measures, as against which
+	// rows its reader may see at all. See populationRule: the zero value
+	// measures the caller's own.
+	population populationRule
 	// referenceScopes names the columns that point at a ROW-SCOPED record of
 	// another kind, mapped to that record's table. The engine's own gate covers
 	// spec.entity — the deals of this report — and says nothing about the
@@ -447,49 +451,3 @@ func percentileFor(fn string) string {
 }
 
 var errUnknownEntity = errors.New("compose: entity outside the schema descriptors")
-
-// runAdHocPlan serves the datasource seam's RunReport: the plan's
-// vocabulary is the schema descriptors (every declared field may group
-// or filter; count is the aggregate). Used by overlay tooling and the
-// seam conformance tests rather than the HTTP surface.
-func (e *reportEngine) runAdHocPlan(ctx context.Context, plan datasource.ReportPlan) (datasource.ReportResult, error) {
-	fields, ok := schemaFields(plan.Entity)
-	if !ok {
-		return datasource.ReportResult{}, errUnknownEntity
-	}
-	spec := reportSpec{
-		entity:       plan.Entity,
-		table:        string(plan.Entity),
-		baseWhere:    whereArchivedNull,
-		activityWalk: plan.Entity == datasource.EntityActivity,
-		dimensions:   map[string]string{},
-		measures:     map[string]string{},
-		filters:      map[string]string{},
-		defaultAggs:  []reportAggregate{{Fn: aggFnCount, As: aliasCount}},
-	}
-	for _, f := range fields {
-		expr := "t." + f.Name
-		spec.dimensions[f.Name] = expr
-		spec.filters[f.Name] = expr
-		if f.Type == "bigint" || f.Type == "integer" {
-			spec.measures[f.Name] = expr
-		}
-	}
-	req := reportRequest{GroupBy: plan.GroupBy, Filters: map[string]any{}}
-	for k, v := range plan.Filter {
-		req.Filters[k] = v
-	}
-	outcome, err := e.runSpec(ctx, "adhoc:"+string(plan.Entity), spec, req)
-	if err != nil {
-		return datasource.ReportResult{}, err
-	}
-	result := datasource.ReportResult{Columns: outcome.Columns}
-	for _, row := range outcome.Rows {
-		values := make([]any, len(outcome.Columns))
-		for i, col := range outcome.Columns {
-			values[i] = row[col]
-		}
-		result.Rows = append(result.Rows, values)
-	}
-	return result, nil
-}
